requirements.txt
Faker==13.6.0
kafka-python==2.0.2
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.2.0
